// src/sqli/dtos/create-customer.dto.ts
export class CreateCustomerDto {
  customerName: string | undefined;
  contactName: string | undefined;
  address: string | undefined;
  city: string | undefined;
  postalCode: string | undefined;
  country: string | undefined;
  username: string | undefined;
  password: string | undefined;
  phoneNumber: string | undefined;
}